requirements.txt
aniso8601==9.0.1
anyio==3.6.2
APScheduler==3.6.3
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
dialogflow==1.1.1
Flask==2.2.3
Flask-RESTful==0.3.9
fonttools==4.39.3
google-api-core==1.34.0
google-auth==2.17.3
google-cloud-dialogflow==2.21.0
googleapis-common-protos==1.59.0
googletrans==4.0.0rc1
grpcio==1.54.0
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
language-tool-python==2.7.1
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclustering==0.10.1.2
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-telegram-bot==13.4.1
python-telegram-bot-raw==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2023.6.3
requests==2.28.2
Reverso-API==0.0.1b3
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scikit-learn-extra==0.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
textblob==0.17.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
translate==3.6.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3
zope.interface==6.0
